xds: Move MockStream to stream_test.go
This commit is to move MockStream struct to stream_test.go as it's only
used in unit test.
